refactor: moved RuleEvaluation
